Resolve an entity description from its morphs before creating the entity from it alone.
